refactor(client): moves header's subcomponents into their own directories
