+++ src/client/src/content/header/components/menu/UserMenuButton.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components/macro";
-import { Button } from "../../../complib/buttons";
+import { Button } from "../../../../complib/buttons";
 
 export const UserMenuButton = styled(Button).attrs(() => ({
   variant: "text",
